+++ traffic_signaling/include/city_plan/street.hpp
@@ -9,12 +9,12 @@ class Intersection;
 
 class Street {
 public:
-    Street(size_t id, const Intersection &start, const Intersection &end,
-           std::string &&name, size_t length)
+    Street(unsigned long id, const Intersection &start, const Intersection &end,
+           std::string &&name, unsigned long length)
         : id_(id), start_(start), end_(end),
           name_(std::move(name)), length_(length) {}
 
-    size_t id() const {
+    unsigned long id() const {
         return id_;
     }
 
@@ -34,7 +34,7 @@ public:
         ++total_cars_;
     }
 
-    size_t total_cars() const {
+    unsigned long total_cars() const {
         return total_cars_;
     }
 
@@ -42,18 +42,18 @@ public:
         return total_cars_ > 0;
     }
 
-    size_t length() const {
+    unsigned long length() const {
         return length_;
     }
 
 private:
-    size_t id_;
+    unsigned long id_;
     const Intersection &start_;
     const Intersection &end_;
     std::string name_;
-    size_t length_;
+    unsigned long length_;
     // number of cars that use this street
-    size_t total_cars_{};
+    unsigned long total_cars_{};
     // TODO: maybe use a vector of cars instead of a counter
 };
 }
